Equipment test cases and related fixes

+++ solaris/warbook/equipment/tonnage.py
@@ -1,4 +1,5 @@
 from math import ceil
+from decimal import Decimal
 
 tonnage_funcs = (
 	('fixed'  , 'Fixed Tonnage'),
@@ -46,7 +47,7 @@ def structure(self, mech=None):
     return ceil((mech.tonnage * self.tonnage_factor * 2) / 10)/2
     
 def gyro(self, mech=None):
-    return ceil(mech.engine_rating / 100.0) * self.tonnage_factor 
+    return Decimal(ceil(mech.engine_rating / 100.0)) * self.tonnage_factor 
    
 def targetting_computer(self, mech=None):
     return ceil(mech.directfire_tonnage() / 4.0)
